# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catering Webapp API",
        "description": "Dokumentasi API untuk sistem pemesanan paket katering (role Penjual & Pelanggan).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pelanggan/pesanan/{pesanan}/review": {
            "post": {
                "tags": [
                    "Pelanggan - Review"
                ],
                "summary": "Beri rating & ulasan untuk pesanan yang sudah selesai",
                "description": "Hanya pemilik pesanan, pesanan harus berstatus \"selesai\", dan 1 pesanan cuma boleh direview sekali.",
                "operationId": "1a8f991508e91b29804badb971100e19",
                "parameters": [
                    {
                        "name": "pesanan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "ulasan": {
                                        "type": "string",
                                        "example": "Makanannya enak, pengiriman tepat waktu!",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review berhasil disimpan"
                    },
                    "403": {
                        "description": "Bukan pemilik pesanan atau pesanan belum selesai"
                    },
                    "422": {
                        "description": "Pesanan ini sudah pernah direview"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/penjual/gubukans": {
            "get": {
                "tags": [
                    "Penjual - Gubukan"
                ],
                "summary": "Daftar semua gubukan",
                "operationId": "4a2e5cc402512e220990434c2fd62007",
                "responses": {
                    "200": {
                        "description": "Daftar gubukan (paginated)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Penjual - Gubukan"
                ],
                "summary": "Tambah gubukan baru",
                "operationId": "e66fe166ffbbb2996f4bc0c354ae7d45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_gubukan",
                                    "harga_gubukan",
                                    "kapasitas_orang"
                                ],
                                "properties": {
                                    "nama_gubukan": {
                                        "type": "string",
                                        "example": "Gubukan Sedang"
                                    },
                                    "harga_gubukan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250000
                                    },
                                    "kapasitas_orang": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "status_aktif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gubukan berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penjual/gubukans/{gubukan}": {
            "get": {
                "tags": [
                    "Penjual - Gubukan"
                ],
                "summary": "Detail 1 gubukan",
                "operationId": "f4ae4c420f9b9b9ccff8d07da6e12550",
                "parameters": [
                    {
                        "name": "gubukan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail gubukan"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Penjual - Gubukan"
                ],
                "summary": "Update gubukan",
                "operationId": "a2762dcb17fb5c0167acefbe8f27399a",
                "parameters": [
                    {
                        "name": "gubukan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_gubukan",
                                    "harga_gubukan",
                                    "kapasitas_orang"
                                ],
                                "properties": {
                                    "nama_gubukan": {
                                        "type": "string"
                                    },
                                    "harga_gubukan": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "kapasitas_orang": {
                                        "type": "integer"
                                    },
                                    "status_aktif": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gubukan berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Penjual - Gubukan"
                ],
                "summary": "Hapus gubukan",
                "description": "Aman dihapus meski sudah pernah dipakai di pesanan lama (gubukan_id di pesanan itu otomatis jadi NULL, pesanannya tidak ikut terhapus).",
                "operationId": "4510e43516647cb940531ecf49e6780b",
                "parameters": [
                    {
                        "name": "gubukan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gubukan berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login dulu lewat /login, lalu isi token yang didapat di sini (tanpa kata \"Bearer\").",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}